+++ plugins/node/instrumentation-fs/src/instrumentation.ts
@@ -144,9 +144,9 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
   ): T {
     const instrumentation = this;
     const patchedFunction = <any>function (this: any, ...args: any[]) {
-      if (isTracingSuppressed(api.context.active())) {
-        // Performance optimization. Avoid creating additional contexts and spans
-        // if we already know that the tracing is being suppressed.
+      const activeContext = api.context.active();
+
+      if (!instrumentation._shouldTrace(activeContext)) {
         return original.apply(this, args);
       }
       if (
@@ -155,7 +155,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         }) === false
       ) {
         return api.context.with(
-          suppressTracing(api.context.active()),
+          suppressTracing(activeContext),
           original,
           this,
           ...args
@@ -169,14 +169,14 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
       try {
         // Suppress tracing for internal fs calls
         const res = api.context.with(
-          suppressTracing(api.trace.setSpan(api.context.active(), span)),
+          suppressTracing(api.trace.setSpan(activeContext, span)),
           original,
           this,
           ...args
         );
         instrumentation._runEndHook(functionName, { args: args, span });
         return res;
-      } catch (error) {
+      } catch (error: any) {
         span.recordException(error);
         span.setStatus({
           message: error.message,
@@ -197,9 +197,9 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
   ): T {
     const instrumentation = this;
     const patchedFunction = <any>function (this: any, ...args: any[]) {
-      if (isTracingSuppressed(api.context.active())) {
-        // Performance optimization. Avoid creating additional contexts and spans
-        // if we already know that the tracing is being suppressed.
+      const activeContext = api.context.active();
+
+      if (!instrumentation._shouldTrace(activeContext)) {
         return original.apply(this, args);
       }
       if (
@@ -208,7 +208,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         }) === false
       ) {
         return api.context.with(
-          suppressTracing(api.context.active()),
+          suppressTracing(activeContext),
           original,
           this,
           ...args
@@ -224,7 +224,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
 
         // Return to the context active during the call in the callback
         args[lastIdx] = api.context.bind(
-          api.context.active(),
+          activeContext,
           function (this: unknown, error?: Error) {
             if (error) {
               span.recordException(error);
@@ -246,12 +246,12 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         try {
           // Suppress tracing for internal fs calls
           return api.context.with(
-            suppressTracing(api.trace.setSpan(api.context.active(), span)),
+            suppressTracing(api.trace.setSpan(activeContext, span)),
             original,
             this,
             ...args
           );
-        } catch (error) {
+        } catch (error: any) {
           span.recordException(error);
           span.setStatus({
             message: error.message,
@@ -278,9 +278,9 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
   >(functionName: 'exists', original: T): T {
     const instrumentation = this;
     const patchedFunction = <any>function (this: any, ...args: any[]) {
-      if (isTracingSuppressed(api.context.active())) {
-        // Performance optimization. Avoid creating additional contexts and spans
-        // if we already know that the tracing is being suppressed.
+      const activeContext = api.context.active();
+
+      if (!instrumentation._shouldTrace(activeContext)) {
         return original.apply(this, args);
       }
       if (
@@ -289,7 +289,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         }) === false
       ) {
         return api.context.with(
-          suppressTracing(api.context.active()),
+          suppressTracing(activeContext),
           original,
           this,
           ...args
@@ -305,7 +305,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
 
         // Return to the context active during the call in the callback
         args[lastIdx] = api.context.bind(
-          api.context.active(),
+          activeContext,
           function (this: unknown) {
             // `exists` never calls the callback with an error
             instrumentation._runEndHook(functionName, {
@@ -320,12 +320,12 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         try {
           // Suppress tracing for internal fs calls
           return api.context.with(
-            suppressTracing(api.trace.setSpan(api.context.active(), span)),
+            suppressTracing(api.trace.setSpan(activeContext, span)),
             original,
             this,
             ...args
           );
-        } catch (error) {
+        } catch (error: any) {
           span.recordException(error);
           span.setStatus({
             message: error.message,
@@ -367,9 +367,9 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
   ): T {
     const instrumentation = this;
     const patchedFunction = <any>async function (this: any, ...args: any[]) {
-      if (isTracingSuppressed(api.context.active())) {
-        // Performance optimization. Avoid creating additional contexts and spans
-        // if we already know that the tracing is being suppressed.
+      const activeContext = api.context.active();
+
+      if (!instrumentation._shouldTrace(activeContext)) {
         return original.apply(this, args);
       }
       if (
@@ -378,7 +378,7 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         }) === false
       ) {
         return api.context.with(
-          suppressTracing(api.context.active()),
+          suppressTracing(activeContext),
           original,
           this,
           ...args
@@ -392,14 +392,14 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
       try {
         // Suppress tracing for internal fs calls
         const res = await api.context.with(
-          suppressTracing(api.trace.setSpan(api.context.active(), span)),
+          suppressTracing(api.trace.setSpan(activeContext, span)),
           original,
           this,
           ...args
         );
         instrumentation._runEndHook(functionName, { args: args, span });
         return res;
-      } catch (error) {
+      } catch (error: any) {
         span.recordException(error);
         span.setStatus({
           message: error.message,
@@ -437,5 +437,23 @@ export default class FsInstrumentation extends InstrumentationBase<FS> {
         this._diag.error('caught endHook error', e);
       }
     }
+  }
+
+  protected _shouldTrace(context: api.Context): boolean {
+    if (isTracingSuppressed(context)) {
+      // Performance optimization. Avoid creating additional contexts and spans
+      // if we already know that the tracing is being suppressed.
+      return false;
+    }
+
+    const { requireParentSpan } = this.getConfig() as FsInstrumentationConfig;
+    if (requireParentSpan) {
+      const parentSpan = api.trace.getSpan(context);
+      if (parentSpan == null) {
+        return false;
+      }
+    }
+
+    return true;
   }
 }