+++ plugins/node/opentelemetry-instrumentation-restify/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.32.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-restify-v0.32.0...instrumentation-restify-v0.32.1) (2023-02-07)
+
+
+### Bug Fixes
+
+* **deps:** update otel core experimental to ^0.35.1 ([#1358](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1358)) ([ff109b7](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/ff109b77928cc9a139a21c63d6b54399bb017fa4))
+
+## [0.32.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-restify-v0.31.0...instrumentation-restify-v0.32.0) (2022-12-20)
+
+
+### Features
+
+* **restify:** add requestHook support ([#1312](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1312)) ([4098e6a](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/4098e6a3d4257e5da9b8cece430bde7d70319cf3))
+
 ## [0.31.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-restify-v0.30.1...instrumentation-restify-v0.31.0) (2022-11-16)